[UPDATE] Nicknamesをachievementパッケージに移動する
