+++ src/main/scala/com/github/unchama/seichiassist/achievement/NicknameMapping.scala
@@ -1,0 +1,192 @@
+package com.github.unchama.seichiassist.achievement
+
+import com.github.unchama.seichiassist.SeichiAssist
+
+object NicknameMapping {
+  case class NicknameCombination(first: Option[AchievementId],
+                                 second: Option[AchievementId] = None,
+                                 third: Option[AchievementId] = None)
+
+  val mapping: Map[AchievementId, NicknameCombination] = Map(
+    1001 -> NicknameCombination(Some(1001)),
+    1002 -> NicknameCombination(Some(1002)),
+    1003 -> NicknameCombination(Some(1003)),
+    1004 -> NicknameCombination(Some(1004)),
+    1005 -> NicknameCombination(Some(1005), None, Some(1005)),
+    1006 -> NicknameCombination(Some(1006)),
+    1007 -> NicknameCombination(Some(1007), Some(9904), Some(1007)),
+    1008 -> NicknameCombination(Some(1008), Some(9901), Some(1008)),
+    1009 -> NicknameCombination(Some(1009), Some(9909), Some(1009)),
+    1010 -> NicknameCombination(Some(1010)),
+    1011 -> NicknameCombination(Some(1011), Some(9904), Some(1011)),
+    1012 -> NicknameCombination(Some(1012), Some(9901), Some(1012)),
+
+    3001 -> NicknameCombination(Some(3001)),
+    3002 -> NicknameCombination(Some(3002), Some(9905), Some(3002)),
+    3003 -> NicknameCombination(Some(3003)),
+    3004 -> NicknameCombination(Some(3004), Some(9902), None),
+    3005 -> NicknameCombination(Some(3005), None, Some(3005)),
+    3006 -> NicknameCombination(Some(3006)),
+    3007 -> NicknameCombination(Some(3007), Some(9905), None),
+    3008 -> NicknameCombination(Some(3008)),
+    3009 -> NicknameCombination(Some(3009), None, Some(3009)),
+    3010 -> NicknameCombination(Some(3010), Some(9909), Some(3010)),
+    3011 -> NicknameCombination(Some(3011)),
+    3012 -> NicknameCombination(Some(3012), None, Some(3012)),
+    3013 -> NicknameCombination(Some(3013), Some(9905), Some(3013)),
+    3014 -> NicknameCombination(Some(3014), Some(9909), Some(3014)),
+    3015 -> NicknameCombination(Some(3015)),
+    3016 -> NicknameCombination(Some(3016)),
+    3017 -> NicknameCombination(Some(3017)),
+    3018 -> NicknameCombination(Some(3018)),
+    3019 -> NicknameCombination(Some(3019)),
+
+    4001 -> NicknameCombination(Some(4001), Some(9905), Some(4001)),
+    4002 -> NicknameCombination(Some(4002), None, Some(4002)),
+    4003 -> NicknameCombination(Some(4003), None, Some(4003)),
+    4004 -> NicknameCombination(Some(4004), Some(9905), Some(4004)),
+    4005 -> NicknameCombination(Some(4005), None, Some(4005)),
+    4006 -> NicknameCombination(Some(4006), Some(9905), Some(4006)),
+    4007 -> NicknameCombination(Some(4007), None, Some(4007)),
+    4008 -> NicknameCombination(Some(4008), None, Some(4008)),
+    4009 -> NicknameCombination(Some(4009), None, Some(4009)),
+    4010 -> NicknameCombination(Some(4010), Some(9905), Some(4010)),
+    4011 -> NicknameCombination(Some(4011), Some(9901), Some(4011)),
+    4012 -> NicknameCombination(Some(4012), None, Some(4012)),
+    4013 -> NicknameCombination(Some(4013), None, Some(4013)),
+    4014 -> NicknameCombination(Some(4014), Some(9905), Some(4014)),
+    4015 -> NicknameCombination(Some(4015)),
+    4016 -> NicknameCombination(Some(4016), None, Some(4016)),
+    4017 -> NicknameCombination(Some(4017)),
+    4018 -> NicknameCombination(Some(4018), None, Some(4018)),
+    4019 -> NicknameCombination(Some(4019), None, Some(4019)),
+    4020 -> NicknameCombination(Some(4020), None, Some(4020)),
+    4021 -> NicknameCombination(Some(4021), None, Some(4021)),
+    4022 -> NicknameCombination(Some(4022), Some(9903), Some(4022)),
+    4023 -> NicknameCombination(Some(4023), None, Some(4023)),
+
+    5001 -> NicknameCombination(Some(5001), Some(5001)),
+    5002 -> NicknameCombination(Some(5002), None, Some(5002)),
+    5003 -> NicknameCombination(Some(5003)),
+    5004 -> NicknameCombination(Some(5004), None, Some(5004)),
+    5005 -> NicknameCombination(Some(5005), None, Some(5005)),
+    5006 -> NicknameCombination(Some(5006), None, Some(5006)),
+    5007 -> NicknameCombination(Some(5007)),
+    5008 -> NicknameCombination(Some(5008), Some(9905)),
+
+    5101 -> NicknameCombination(Some(5101), None, Some(5010)),
+    5102 -> NicknameCombination(Some(5102), Some(9907), Some(5102)),
+    5103 -> NicknameCombination(Some(5103), Some(9905), None),
+    5104 -> NicknameCombination(Some(5104), Some(9905), Some(5104)),
+    5105 -> NicknameCombination(Some(5105), Some(9907), Some(5105)),
+    5106 -> NicknameCombination(Some(5106)),
+    5107 -> NicknameCombination(Some(5107), Some(9909), Some(5107)),
+    5108 -> NicknameCombination(Some(5108)),
+    5109 -> NicknameCombination(Some(5109)),
+    5110 -> NicknameCombination(Some(5110)),
+    5111 -> NicknameCombination(Some(5111)),
+    5112 -> NicknameCombination(Some(5112), None, Some(5112)),
+    5113 -> NicknameCombination(Some(5113), Some(9905), Some(5113)),
+    5114 -> NicknameCombination(Some(5114), None, Some(5114)),
+    5115 -> NicknameCombination(Some(5115)),
+    5116 -> NicknameCombination(Some(5116), Some(9905), Some(5116)),
+    5117 -> NicknameCombination(Some(5117), None, Some(5117)),
+    5118 -> NicknameCombination(Some(5118), None, Some(5118)),
+    5119 -> NicknameCombination(Some(5119), Some(9905), Some(5119)),
+    5120 -> NicknameCombination(Some(5120), Some(5120), Some(5120)),
+
+    6001 -> NicknameCombination(Some(6001)),
+    6002 -> NicknameCombination(Some(6002), None, Some(6002)),
+    6003 -> NicknameCombination(Some(6003)),
+    6004 -> NicknameCombination(Some(6004), Some(9903), Some(6004)),
+    6005 -> NicknameCombination(Some(6005), Some(9905)),
+    6006 -> NicknameCombination(Some(6006), None, Some(6006)),
+    6007 -> NicknameCombination(Some(6007), Some(9902)),
+    6008 -> NicknameCombination(Some(6008)),
+
+    7001 -> NicknameCombination(Some(7001), Some(9901), Some(7001)),
+    7002 -> NicknameCombination(Some(7002), Some(9905), Some(7002)),
+    7003 -> NicknameCombination(Some(7003), Some(9905), Some(7003)),
+    7004 -> NicknameCombination(None, Some(7004), None),
+    7005 -> NicknameCombination(Some(7005), Some(9902), Some(7005)),
+    7006 -> NicknameCombination(Some(7006), Some(9905), Some(7006)),
+    7007 -> NicknameCombination(Some(7007), Some(9905), Some(7007)),
+    7008 -> NicknameCombination(Some(7008), Some(9905), Some(7008)),
+    7009 -> NicknameCombination(Some(7009), Some(9905), Some(7009)),
+    7010 -> NicknameCombination(Some(7010), None, Some(7010)),
+    7011 -> NicknameCombination(Some(7011), Some(9905), Some(7011)),
+    7012 -> NicknameCombination(Some(7012), None, Some(7012)),
+    7013 -> NicknameCombination(Some(7013)),
+    7014 -> NicknameCombination(Some(7014)),
+    7015 -> NicknameCombination(Some(7015), Some(9904), Some(7015)),
+    7016 -> NicknameCombination(Some(7016), None, Some(7016)),
+    7017 -> NicknameCombination(Some(7017), Some(9905), Some(7017)),
+    7018 -> NicknameCombination(Some(7018), Some(9904), Some(7018)),
+    7019 -> NicknameCombination(Some(7019), None, Some(7019)),
+    7020 -> NicknameCombination(Some(7020), None, Some(7020)),
+    7021 -> NicknameCombination(Some(7021), Some(9905), Some(7021)),
+    7022 -> NicknameCombination(Some(7022), None, Some(7022)),
+    7023 -> NicknameCombination(Some(7023), Some(9905), Some(7023)),
+    7024 -> NicknameCombination(Some(7024), None, Some(7024)),
+    7025 -> NicknameCombination(Some(7025), Some(9905), Some(7025)),
+    7026 -> NicknameCombination(Some(7026), Some(9905), Some(7025)),
+    7027 -> NicknameCombination(Some(7027), None, Some(7027)),
+
+    7901 -> NicknameCombination(Some(7901), Some(7901), Some(7910)),
+    7902 -> NicknameCombination(Some(7902), None, Some(7902)),
+    7903 -> NicknameCombination(Some(7903), Some(9905), Some(7903)),
+    7904 -> NicknameCombination(Some(7904), Some(9907), Some(7904)),
+    7905 -> NicknameCombination(Some(7905), None, Some(7905)),
+    7906 -> NicknameCombination(Some(7906), None, Some(7906)),
+
+    9001 -> NicknameCombination(Some(9001)),
+    9002 -> NicknameCombination(Some(9002), None, Some(9002)),
+    9003 -> NicknameCombination(Some(9003)),
+    9004 -> NicknameCombination(Some(9004), Some(9004), Some(9004)),
+    9005 -> NicknameCombination(Some(9005), None, Some(9005)),
+    9006 -> NicknameCombination(Some(9006)),
+    9007 -> NicknameCombination(Some(9007)),
+    9008 -> NicknameCombination(Some(9008), None, Some(9008)),
+    9009 -> NicknameCombination(Some(9009)),
+    9010 -> NicknameCombination(Some(9010), Some(9903), Some(9010)),
+    9011 -> NicknameCombination(Some(9011), None, Some(9011)),
+    9012 -> NicknameCombination(Some(9012), None, Some(9012)),
+    9013 -> NicknameCombination(Some(9013)),
+    9014 -> NicknameCombination(None, Some(9014), None),
+    9015 -> NicknameCombination(Some(9015), None, Some(9015)),
+    9016 -> NicknameCombination(Some(9016), Some(9016), Some(9016)),
+    9017 -> NicknameCombination(Some(9017), None, Some(9017)),
+    9018 -> NicknameCombination(Some(9018)),
+    9019 -> NicknameCombination(Some(9019), Some(9901), Some(9019)),
+    9020 -> NicknameCombination(Some(9020), None, Some(9020)),
+    9021 -> NicknameCombination(Some(9021), Some(9901), Some(9021)),
+    9022 -> NicknameCombination(Some(9022), None, Some(9022)),
+    9023 -> NicknameCombination(Some(9023), None, Some(9023)),
+    9024 -> NicknameCombination(Some(9024), None, Some(9024)),
+    9025 -> NicknameCombination(Some(9025), Some(9025), Some(9025)),
+    9026 -> NicknameCombination(Some(9026), None, Some(9026)),
+    9027 -> NicknameCombination(Some(9027), None, Some(9027)),
+    9028 -> NicknameCombination(Some(9028), Some(9028), Some(9028)),
+    9029 -> NicknameCombination(Some(9029), Some(9029), Some(9029)),
+    9030 -> NicknameCombination(Some(9030), Some(9905), Some(9030)),
+    9031 -> NicknameCombination(Some(9031), Some(9908), Some(9031)),
+    9032 -> NicknameCombination(Some(9032), None, Some(9032)),
+    9033 -> NicknameCombination(Some(9033), Some(9903), Some(9033)),
+    9034 -> NicknameCombination(Some(9034), None, Some(9034)),
+    9035 -> NicknameCombination(Some(9035), Some(9905), Some(9035)),
+    9036 -> NicknameCombination(Some(9036), None, Some(9036)),
+
+    8001 -> NicknameCombination(Some(8001), Some(9905), Some(8001)),
+    8002 -> NicknameCombination(Some(8002), Some(9905), Some(8002)),
+    8003 -> NicknameCombination(Some(8003), None, Some(8003)),
+  )
+
+  def getTitleFor(id: AchievementId): Option[String] =
+    mapping.get(id).map { case NicknameCombination(first, second, third) =>
+      val config = SeichiAssist.seichiAssistConfig
+
+      first.map(config.getTitle1).getOrElse("") +
+        second.map(config.getTitle2).getOrElse("") +
+        third.map(config.getTitle3).getOrElse("")
+    }
+}
