--- src/main/scala/com/github/unchama/seichiassist/achievement/TitleMapping.scala
@@ -1,192 +1,0 @@
-package com.github.unchama.seichiassist.achievement
-
-import com.github.unchama.seichiassist.SeichiAssist
-
-object TitleMapping {
-  case class TitleCombination(first: Option[AchievementId],
-                              second: Option[AchievementId] = None,
-                              third: Option[AchievementId] = None)
-
-  val mapping: Map[AchievementId, TitleCombination] = Map(
-    1001 -> TitleCombination(Some(1001)),
-    1002 -> TitleCombination(Some(1002)),
-    1003 -> TitleCombination(Some(1003)),
-    1004 -> TitleCombination(Some(1004)),
-    1005 -> TitleCombination(Some(1005), None, Some(1005)),
-    1006 -> TitleCombination(Some(1006)),
-    1007 -> TitleCombination(Some(1007), Some(9904), Some(1007)),
-    1008 -> TitleCombination(Some(1008), Some(9901), Some(1008)),
-    1009 -> TitleCombination(Some(1009), Some(9909), Some(1009)),
-    1010 -> TitleCombination(Some(1010)),
-    1011 -> TitleCombination(Some(1011), Some(9904), Some(1011)),
-    1012 -> TitleCombination(Some(1012), Some(9901), Some(1012)),
-
-    3001 -> TitleCombination(Some(3001)),
-    3002 -> TitleCombination(Some(3002), Some(9905), Some(3002)),
-    3003 -> TitleCombination(Some(3003)),
-    3004 -> TitleCombination(Some(3004), Some(9902), None),
-    3005 -> TitleCombination(Some(3005), None, Some(3005)),
-    3006 -> TitleCombination(Some(3006)),
-    3007 -> TitleCombination(Some(3007), Some(9905), None),
-    3008 -> TitleCombination(Some(3008)),
-    3009 -> TitleCombination(Some(3009), None, Some(3009)),
-    3010 -> TitleCombination(Some(3010), Some(9909), Some(3010)),
-    3011 -> TitleCombination(Some(3011)),
-    3012 -> TitleCombination(Some(3012), None, Some(3012)),
-    3013 -> TitleCombination(Some(3013), Some(9905), Some(3013)),
-    3014 -> TitleCombination(Some(3014), Some(9909), Some(3014)),
-    3015 -> TitleCombination(Some(3015)),
-    3016 -> TitleCombination(Some(3016)),
-    3017 -> TitleCombination(Some(3017)),
-    3018 -> TitleCombination(Some(3018)),
-    3019 -> TitleCombination(Some(3019)),
-
-    4001 -> TitleCombination(Some(4001), Some(9905), Some(4001)),
-    4002 -> TitleCombination(Some(4002), None, Some(4002)),
-    4003 -> TitleCombination(Some(4003), None, Some(4003)),
-    4004 -> TitleCombination(Some(4004), Some(9905), Some(4004)),
-    4005 -> TitleCombination(Some(4005), None, Some(4005)),
-    4006 -> TitleCombination(Some(4006), Some(9905), Some(4006)),
-    4007 -> TitleCombination(Some(4007), None, Some(4007)),
-    4008 -> TitleCombination(Some(4008), None, Some(4008)),
-    4009 -> TitleCombination(Some(4009), None, Some(4009)),
-    4010 -> TitleCombination(Some(4010), Some(9905), Some(4010)),
-    4011 -> TitleCombination(Some(4011), Some(9901), Some(4011)),
-    4012 -> TitleCombination(Some(4012), None, Some(4012)),
-    4013 -> TitleCombination(Some(4013), None, Some(4013)),
-    4014 -> TitleCombination(Some(4014), Some(9905), Some(4014)),
-    4015 -> TitleCombination(Some(4015)),
-    4016 -> TitleCombination(Some(4016), None, Some(4016)),
-    4017 -> TitleCombination(Some(4017)),
-    4018 -> TitleCombination(Some(4018), None, Some(4018)),
-    4019 -> TitleCombination(Some(4019), None, Some(4019)),
-    4020 -> TitleCombination(Some(4020), None, Some(4020)),
-    4021 -> TitleCombination(Some(4021), None, Some(4021)),
-    4022 -> TitleCombination(Some(4022), Some(9903), Some(4022)),
-    4023 -> TitleCombination(Some(4023), None, Some(4023)),
-
-    5001 -> TitleCombination(Some(5001), Some(5001)),
-    5002 -> TitleCombination(Some(5002), None, Some(5002)),
-    5003 -> TitleCombination(Some(5003)),
-    5004 -> TitleCombination(Some(5004), None, Some(5004)),
-    5005 -> TitleCombination(Some(5005), None, Some(5005)),
-    5006 -> TitleCombination(Some(5006), None, Some(5006)),
-    5007 -> TitleCombination(Some(5007)),
-    5008 -> TitleCombination(Some(5008), Some(9905)),
-
-    5101 -> TitleCombination(Some(5101), None, Some(5010)),
-    5102 -> TitleCombination(Some(5102), Some(9907), Some(5102)),
-    5103 -> TitleCombination(Some(5103), Some(9905), None),
-    5104 -> TitleCombination(Some(5104), Some(9905), Some(5104)),
-    5105 -> TitleCombination(Some(5105), Some(9907), Some(5105)),
-    5106 -> TitleCombination(Some(5106)),
-    5107 -> TitleCombination(Some(5107), Some(9909), Some(5107)),
-    5108 -> TitleCombination(Some(5108)),
-    5109 -> TitleCombination(Some(5109)),
-    5110 -> TitleCombination(Some(5110)),
-    5111 -> TitleCombination(Some(5111)),
-    5112 -> TitleCombination(Some(5112), None, Some(5112)),
-    5113 -> TitleCombination(Some(5113), Some(9905), Some(5113)),
-    5114 -> TitleCombination(Some(5114), None, Some(5114)),
-    5115 -> TitleCombination(Some(5115)),
-    5116 -> TitleCombination(Some(5116), Some(9905), Some(5116)),
-    5117 -> TitleCombination(Some(5117), None, Some(5117)),
-    5118 -> TitleCombination(Some(5118), None, Some(5118)),
-    5119 -> TitleCombination(Some(5119), Some(9905), Some(5119)),
-    5120 -> TitleCombination(Some(5120), Some(5120), Some(5120)),
-
-    6001 -> TitleCombination(Some(6001)),
-    6002 -> TitleCombination(Some(6002), None, Some(6002)),
-    6003 -> TitleCombination(Some(6003)),
-    6004 -> TitleCombination(Some(6004), Some(9903), Some(6004)),
-    6005 -> TitleCombination(Some(6005), Some(9905)),
-    6006 -> TitleCombination(Some(6006), None, Some(6006)),
-    6007 -> TitleCombination(Some(6007), Some(9902)),
-    6008 -> TitleCombination(Some(6008)),
-
-    7001 -> TitleCombination(Some(7001), Some(9901), Some(7001)),
-    7002 -> TitleCombination(Some(7002), Some(9905), Some(7002)),
-    7003 -> TitleCombination(Some(7003), Some(9905), Some(7003)),
-    7004 -> TitleCombination(None, Some(7004), None),
-    7005 -> TitleCombination(Some(7005), Some(9902), Some(7005)),
-    7006 -> TitleCombination(Some(7006), Some(9905), Some(7006)),
-    7007 -> TitleCombination(Some(7007), Some(9905), Some(7007)),
-    7008 -> TitleCombination(Some(7008), Some(9905), Some(7008)),
-    7009 -> TitleCombination(Some(7009), Some(9905), Some(7009)),
-    7010 -> TitleCombination(Some(7010), None, Some(7010)),
-    7011 -> TitleCombination(Some(7011), Some(9905), Some(7011)),
-    7012 -> TitleCombination(Some(7012), None, Some(7012)),
-    7013 -> TitleCombination(Some(7013)),
-    7014 -> TitleCombination(Some(7014)),
-    7015 -> TitleCombination(Some(7015), Some(9904), Some(7015)),
-    7016 -> TitleCombination(Some(7016), None, Some(7016)),
-    7017 -> TitleCombination(Some(7017), Some(9905), Some(7017)),
-    7018 -> TitleCombination(Some(7018), Some(9904), Some(7018)),
-    7019 -> TitleCombination(Some(7019), None, Some(7019)),
-    7020 -> TitleCombination(Some(7020), None, Some(7020)),
-    7021 -> TitleCombination(Some(7021), Some(9905), Some(7021)),
-    7022 -> TitleCombination(Some(7022), None, Some(7022)),
-    7023 -> TitleCombination(Some(7023), Some(9905), Some(7023)),
-    7024 -> TitleCombination(Some(7024), None, Some(7024)),
-    7025 -> TitleCombination(Some(7025), Some(9905), Some(7025)),
-    7026 -> TitleCombination(Some(7026), Some(9905), Some(7025)),
-    7027 -> TitleCombination(Some(7027), None, Some(7027)),
-
-    7901 -> TitleCombination(Some(7901), Some(7901), Some(7910)),
-    7902 -> TitleCombination(Some(7902), None, Some(7902)),
-    7903 -> TitleCombination(Some(7903), Some(9905), Some(7903)),
-    7904 -> TitleCombination(Some(7904), Some(9907), Some(7904)),
-    7905 -> TitleCombination(Some(7905), None, Some(7905)),
-    7906 -> TitleCombination(Some(7906), None, Some(7906)),
-
-    9001 -> TitleCombination(Some(9001)),
-    9002 -> TitleCombination(Some(9002), None, Some(9002)),
-    9003 -> TitleCombination(Some(9003)),
-    9004 -> TitleCombination(Some(9004), Some(9004), Some(9004)),
-    9005 -> TitleCombination(Some(9005), None, Some(9005)),
-    9006 -> TitleCombination(Some(9006)),
-    9007 -> TitleCombination(Some(9007)),
-    9008 -> TitleCombination(Some(9008), None, Some(9008)),
-    9009 -> TitleCombination(Some(9009)),
-    9010 -> TitleCombination(Some(9010), Some(9903), Some(9010)),
-    9011 -> TitleCombination(Some(9011), None, Some(9011)),
-    9012 -> TitleCombination(Some(9012), None, Some(9012)),
-    9013 -> TitleCombination(Some(9013)),
-    9014 -> TitleCombination(None, Some(9014), None),
-    9015 -> TitleCombination(Some(9015), None, Some(9015)),
-    9016 -> TitleCombination(Some(9016), Some(9016), Some(9016)),
-    9017 -> TitleCombination(Some(9017), None, Some(9017)),
-    9018 -> TitleCombination(Some(9018)),
-    9019 -> TitleCombination(Some(9019), Some(9901), Some(9019)),
-    9020 -> TitleCombination(Some(9020), None, Some(9020)),
-    9021 -> TitleCombination(Some(9021), Some(9901), Some(9021)),
-    9022 -> TitleCombination(Some(9022), None, Some(9022)),
-    9023 -> TitleCombination(Some(9023), None, Some(9023)),
-    9024 -> TitleCombination(Some(9024), None, Some(9024)),
-    9025 -> TitleCombination(Some(9025), Some(9025), Some(9025)),
-    9026 -> TitleCombination(Some(9026), None, Some(9026)),
-    9027 -> TitleCombination(Some(9027), None, Some(9027)),
-    9028 -> TitleCombination(Some(9028), Some(9028), Some(9028)),
-    9029 -> TitleCombination(Some(9029), Some(9029), Some(9029)),
-    9030 -> TitleCombination(Some(9030), Some(9905), Some(9030)),
-    9031 -> TitleCombination(Some(9031), Some(9908), Some(9031)),
-    9032 -> TitleCombination(Some(9032), None, Some(9032)),
-    9033 -> TitleCombination(Some(9033), Some(9903), Some(9033)),
-    9034 -> TitleCombination(Some(9034), None, Some(9034)),
-    9035 -> TitleCombination(Some(9035), Some(9905), Some(9035)),
-    9036 -> TitleCombination(Some(9036), None, Some(9036)),
-
-    8001 -> TitleCombination(Some(8001), Some(9905), Some(8001)),
-    8002 -> TitleCombination(Some(8002), Some(9905), Some(8002)),
-    8003 -> TitleCombination(Some(8003), None, Some(8003)),
-  )
-
-  def getTitleFor(id: AchievementId): Option[String] =
-    mapping.get(id).map { case TitleCombination(first, second, third) =>
-      val config = SeichiAssist.seichiAssistConfig
-
-      first.map(config.getTitle1).getOrElse("") +
-        second.map(config.getTitle2).getOrElse("") +
-        third.map(config.getTitle3).getOrElse("")
-    }
-}
